-Removed some warnings from vjDebug.cpp -Added #define statements for debug levels in vjDebug.h
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@1558 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ juggler_1.0/Kernel/vjDebug.cpp
@@ -87,7 +87,7 @@ void vjDebug::addCategoryName(std::string name, int cat)
 
 void vjDebug::addAllowedCategory(int cat)
 {
-   if(mAllowedCategories.size() < (cat+1))
+   if((int)mAllowedCategories.size() < (cat+1))
       growAllowedCategoryVector(cat+1);
 
    mAllowedCategories[cat] = true;
@@ -151,7 +151,7 @@ void vjDebug::getAllowedCatsFromEnv()
 
 void vjDebug::growAllowedCategoryVector(int newSize)
 {
-   while(mAllowedCategories.size() < newSize)
+   while((int)mAllowedCategories.size() < newSize)
       mAllowedCategories.push_back(false);
 }
